refactor(all): move TransactionDetailsProxy to separate file
